+++ src/components/UiButtonHamburger/index.js
@@ -1,0 +1,4 @@
+import factoryComponentPlugin from '@/utils/factoryComponentPlugin';
+import UiButtonHamburger from './UiButtonHamburger';
+
+export default factoryComponentPlugin(UiButtonHamburger);