+++ src/components/index.js
@@ -1,3 +1,8 @@
+// Grid
+export { default as UiContainer } from './UiContainer';
+export { default as UiRow } from './UiRow';
+export { default as UiCol } from './UiCol';
+
 // Typography
 export { default as UiIcon } from './UiIcon';
 export { default as UiLink } from './UiLink';
@@ -6,6 +11,7 @@ export { default as UiText } from './UiText';
 // Buttons
 export { default as UiButton } from './UiButton';
 export { default as UiButtonGroup } from './UiButtonGroup';
+export { default as UiButtonHamburger } from './UiButtonHamburger';
 
 // Forms
 export { default as UiInput } from './UiInput';
@@ -16,11 +22,6 @@ export { default as UiTextarea } from './UiTextarea';
 /**
  * @TODO: реализовать
  */
-
-// Grid
-export { default as UiContainer } from './UiContainer';
-export { default as UiRow } from './UiRow';
-export { default as UiCol } from './UiCol';
 
 // Other
 export { default as UiApp } from './UiApp';